requirements.txt
argparse
pandas
psycopg2-binary
python-dotenv
pytest
